+++ test/support/test_cache.ex
@@ -119,4 +119,24 @@ defmodule Nebulex.Distributed.TestCache do
         adapter: Nebulex.Adapters.Nil
     end
   end
+
+  defmodule CoherentCache do
+    @moduledoc false
+    use Nebulex.Cache,
+      otp_app: :nebulex_distributed,
+      adapter: Nebulex.Adapters.Coherent,
+      adapter_opts: [primary_storage_adapter: Nebulex.Adapters.Local]
+
+    use Commons
+  end
+
+  defmodule CoherentNilCache do
+    @moduledoc false
+    use Nebulex.Cache,
+      otp_app: :nebulex_distributed,
+      adapter: Nebulex.Adapters.Coherent,
+      adapter_opts: [primary_storage_adapter: Nebulex.Adapters.Nil]
+
+    use Commons
+  end
 end